Add Ruby next wrapper
RubyGems doesn't like non-Ruby executables in your gems: https://github.com/rubygems/rubygems/issues/88

To get around this, I've added a Ruby wrapper. This is slower than adding the script directly into your `bin` directory but it's fine for the purposes of this particular gem. If this were a more permanent solution, I'd be tempted to copy the script to the project's `bin` directory.