Add Ruby 3.4 support
Ruby 3.4 is removing the `base64` default gem to [became bundled](https://rubyreferences.github.io/rubychanges/3.4.html#default-gems-that-became-bundled)

Making the current webmock (v3.16.2) gem incompatible because it has a dependency on the `base64` gem,
to fix it we are bumping webmock to v3.20.0 which [removed the
`base64`](https://github.com/bblimke/webmock/pull/1046) dependency.

> [!NOTE]
> As an alternative solution, we can leave the current `webmock` version untoched
and add the `base64` gem as a dependency.